refactor(web-ui): decompose PortfolioTable god-component
Split the 653-line PortfolioTable into three focused files:
- PortfolioTableParts.tsx: presentational pieces (TimeStopBadge,
  ActionsDropdown, DropdownItem, StopPreviewPanel) + row type + formatters
- portfolioColumns.tsx: buildPortfolioColumns factory taking semantic
  action callbacks
- PortfolioTable.tsx: slim container (state, mutations, modals, row
  derivation), now 348 lines

The cancel-order window.confirm moves from the column render into the
container onCancelOrder handler. No behavior or public interface change.
